chore: move lib/types/base to lib/jhipster/types
