+++ generators/bootstrap-application-base/support/export-jdl-transform.ts
@@ -8,7 +8,7 @@ import { GENERATOR_JHIPSTER } from '../../generator-constants.js';
 import { getJDLObjectFromSingleApplication } from '../../../lib/jdl/converters/json-to-jdl-converter.js';
 import { createRuntime } from '../../../lib/jdl/core/runtime.js';
 import type { JDLApplicationConfig } from '../../../lib/jdl/core/types/parsing.js';
-import type { Entity } from '../../../lib/types/base/entity.js';
+import type { Entity } from '../../../lib/jhipster/types/entity.js';
 
 export const exportJDLTransform = ({
   destinationPath,
