+++ lib/jhipster/types/entity.d.ts
