+++ lib/jhipster/types/property.d.ts
