+++ lib/jhipster/types/relationship.d.ts
